fix: spacing between author and content
